+++ rpcdaemon/plugins/l3agent.py
@@ -22,7 +22,7 @@ class L3Agent(NeutronAgent, RPC):
         self.config = config.section('L3Agent')
 
         # grab relevant settings
-        queue_expire = self.config.get('queue_expire', 60)
+        queue_expire = int(self.config.get('queue_expire', 60))
 
         # Initialize logger
         self.logger = Logger(
@@ -52,7 +52,7 @@ class L3Agent(NeutronAgent, RPC):
                 'durable': False,
                 'routing_key': 'q-plugin',
                 'queue_arguments': {
-                    'x-expires': int(queue_expire * 1000),
+                    'x-expires': queue_expire * 1000,
                 }
             }
         )
